test(langevals): split the credential isolation tests by what they prove
The module had grown past the 300 source line limit, and it held two kinds of
test. The resolution layer tests stay in test_credential_isolation.py: request
env in, litellm call arguments out, one assertion per rule. The end-to-end
sweep over every evaluator family that used to write os.environ moves to
test_evaluator_credential_isolation.py, with the recording provider stub and
the case table it needs. 129 and 247 source lines.

The evaluator module repeats the server import guard the way the other modules
in this directory do, and rebases server.original_env after it. Same 24 tests
pass, and the keyless suite stays at 66.

Also drops the pytest import the resolution module no longer uses.
